requirements.txt
pymongo>=3.10.1<4
gunicorn<20
funcsigs==0.4
gevent==1.1.1
greenlet==0.4.9
iso8601==0.1.11
LEPL==5.1.3
MarkupSafe==0.23
PasteDeploy==1.5.2
pyprof2calltree==1.3.2
pyramid==1.7
pyramid-exclog==0.7
repoze.lru==0.6
repoze.profile==2.1
six==1.10.0
translationstring==1.3
venusian==1.0
wrapt==1.10.8
zope.deprecation==4.1.1
zope.interface==4.1.1
schematics==1.1.1
zope.configuration==4.0.3
zope.component==4.1.0
zope.schema==4.4.2
zope.i18nmessageid==4.0.3
zope.event==4.0.3
pycrypto==2.6.1
jsonpointer==1.9
iowait==0.2
pbkdf2==1.3
tzlocal==1.2.1
py==1.8.0
simplejson==3.6.5
psutil==2.1.3
PyNaCl<2
pyzmq==14.4.1
retrying==1.3.3
beautifulsoup4<5
iso8601==0.1.11
waitress==0.8.10
isodate==0.6.0
python-json-logger==0.1.10
jmespath==0.9.3
execnet==1.1

git+git://github.com/ProzorroUKR/openprocurement.client.python@2.1.1dp#egg=openprocurement_client
git+git://github.com/ProzorroUKR/request_id_middleware.git@0.1.2
git+git://github.com/ProzorroUKR/openprocurement.subscribers.serverid.git@1.2
git+git://github.com/ProzorroUKR/esculator.git@0.0.2
git+git://github.com/ProzorroUKR/dateorro.git@0.0.1
git+git://github.com/ProzorroUKR/rfc6266.git@0.0.6
git+git://github.com/ProzorroUKR/python-json-patch.git@1.14-dp
git+git://github.com/ProzorroUKR/cornice.git@1.2.0.dev0
git+git://github.com/ProzorroUKR/barbecue.git@0.2
git+git://github.com/ProzorroUKR/server_cookie_middleware.git@0.2